+++ app/src/main/AndroidManifest.xml
@@ -9,7 +9,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".WriteActibity"></activity>
+        <activity android:name=".WriteActibity" />
         <activity
             android:name=".MainActivity"
             android:label="@string/app_name">
