+++ app/src/main/java/com/example/sasohan/Fragment2.java
@@ -18,7 +18,7 @@ public class Fragment2 extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_fragment2, container, false);
+        return inflater.inflate(R.layout.sojunghae_1, container, false);
     }
 
 }
